backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
cryptography>=41.0.0
gunicorn==21.2.0
email-validator==2.1.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
